requirements.txt
pytest >= 6.1.1
allure-pytest>= 2.8.40
allure-python-commons>=2.8.40
Faker == 8.2.0
openpyxl == 3.0.7
requests == 2.25.1
python-dotenv == 0.18.0
jsonpath-rw == 1.4.0